refactor: 単一責任となるようにUIの責務でコンポーネントを分割
